requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.6
django-crispy-forms==1.9.2
django-filter==2.3.0
django-heroku==0.3.1
django-progressive-web-app==0.1.1
djangorestframework==3.11.1
gunicorn==20.0.4
idna==2.10
psycopg2==2.8.6
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
stripe==2.51.0
urllib3==1.25.10
whitenoise==5.2.0
